# openapi.yaml
openapi: 3.0.3
info:
  title: Librus Proxy API
  description: Proste API proxy do librus-api (oceny, plan, wiadomości, zadania).
  version: "0.1.0"
servers:
  - url: /
    description: Relative to current host
security:
  - XApiKey: []       # wariant 1: nagłówek
  - QueryApiKey: []   # wariant 2: query param
tags:
  - name: Grades
  - name: Timetable
  - name: Messages
  - name: Homeworks
  - name: Misc
paths:
  /grades:
    get:
      tags: [Grades]
      summary: Pobierz oceny (tylko niepuste przedmioty)
      parameters:
        - in: query
          name: subject
          schema: { type: string }
          description: Filtrowanie po fragmencie nazwy przedmiotu (case-insensitive)
        - in: query
          name: includeRaw
          schema: { type: boolean, default: false }
          description: Dołącz surowe rekordy z librus-api (debug)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradesResponse"

  /debug/grades:
    get:
      tags: [Grades]
      summary: Surowe dane z client.info.getGrades()
      responses:
        "200":
          description: OK

  /timetable:
    get:
      tags: [Timetable]
      summary: Pobierz plan lekcji (z siatki lub z kalendarza)
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
          description: Konkretny dzień (YYYY-MM-DD)
        - in: query
          name: weekStart
          schema: { type: string, format: date }
          description: Poniedziałek tygodnia (mapowanie Monday..Sunday -> daty)
        - in: query
          name: from
          schema: { type: string, format: date }
          description: Początek zakresu (alternatywa dla weekStart)
        - in: query
          name: to
          schema: { type: string, format: date }
          description: Koniec zakresu
        - in: query
          name: includeRaw
          schema: { type: boolean, default: false }
          description: Dołącz surowe rekordy (debug)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableResponse"
  /timetable/today:
    get:
      tags: [Timetable]
      summary: Plan lekcji tylko na dzisiaj
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableTodayResponse"

  /debug/timetable:
    get:
      tags: [Timetable]
      summary: Surowe dane z client.calendar.getTimetable()
      responses:
        "200":
          description: OK

  /debug/calendar:
    get:
      tags: [Timetable]
      summary: Surowe dane z client.calendar.getCalendar()
      responses:
        "200":
          description: OK

  /messages:
    get:
      tags: [Messages]
      summary: Lista wiadomości z folderu
      parameters:
        - in: query
          name: folderId
          schema: { type: integer, default: 5 }
          description: ID folderu (np. 5=Odebrane, 6=Wysłane, 10=Uwagi)
        - in: query
          name: page
          schema: { type: integer }
          description: Numer strony (jeśli wspierane)
        - in: query
          name: includeRaw
          schema: { type: boolean, default: false }
          description: Dołącz surowe rekordy (debug)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesListResponse"

  /messages/{folderId}/{id}:
    get:
      tags: [Messages]
      summary: Szczegóły jednej wiadomości
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: integer }
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: includeRaw
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetailResponse"
        "404":
          description: Nie znaleziono

  /messages/receivers:
    get:
      tags: [Messages]
      summary: Wyszukaj potencjalnych adresatów
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Fraza do wyszukania
      responses:
        "200":
          description: OK

  /messages/feed:
    get:
      tags: [Messages]
      summary: "Wizualny podgląd wiadomości (feed HTML)"
      description: >
        Zwraca najnowsze wiadomości w formacie **HTML**, gotowe do osadzenia np. w
        panelu `iframe` w Home Assistant lub jako prosty mikroblog.  
        Endpoint wymaga autoryzacji (`X-API-Key` w nagłówku lub `?apiKey=` w URL).  
        Można ograniczyć liczbę wiadomości i wybrać folder Librusa (domyślnie 5 = Odebrane).

        💡 Przykładowe użycie w Home Assistant:

        ```yaml
        panel_iframe:
          librus_messages:
            title: "📬 Wiadomości Librus"
            icon: mdi:email
            url: "http://192.168.1.123:3300/messages/feed?apiKey=TWÓJ_KLUCZ"
        ```

      parameters:
        - name: folderId
          in: query
          required: false
          description: "ID folderu Librusa (np. 5 = Odebrane, 10 = Uwagi)"
          schema:
            type: integer
            default: 5
        - name: limit
          in: query
          required: false
          description: "Ilość wiadomości do wyświetlenia (max 200)"
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: "Lista wiadomości w formacie HTML"
          content:
            text/html:
              schema:
                type: string
                example: |
                  <!doctype html>
                  <html>
                  <body>
                    <header><h1>📥 Ostatnie 10 wiadomości</h1></header>
                    <main>
                      <article class="msg">
                        <h2>Informacja dla rodziców</h2>
                        <div class="meta">Bruchajzer Klaudia • 2025-11-07</div>
                        <div class="body">Prosimy o przyniesienie podręczników...</div>
                      </article>
                    </main>
                  </body>
                  </html>
        '401':
          description: "Brak lub błędny klucz API"
        '500':
          description: "Błąd wewnętrzny serwera"

  /messages/feed-json:
    get:
      tags: [Messages]
      summary: "Lista wiadomości (JSON) do użycia w HA"
      description: >
        Zwraca ostatnie wiadomości jako JSON (najświeższe na górze).
        Wymaga autoryzacji (nagłówek `X-API-Key` lub parametr `?apiKey=`).
      parameters:
        - name: apiKey
          in: query
          required: false
          description: "Klucz API (alternatywa dla nagłówka `X-API-Key`)."
          schema: { type: string }
        - name: folderId
          in: query
          required: false
          description: "ID folderu (5 = Odebrane, 10 = Uwagi)."
          schema: { type: integer, default: 5 }
        - name: limit
          in: query
          required: false
          description: "Liczba wiadomości (max 200)."
          schema: { type: integer, default: 50 }
        - name: unreadOnly
          in: query
          required: false
          description: "Tylko nieprzeczytane (true/false)."
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesFeedResponse"
        "401":
          description: Brak / błędny klucz API
        "500":
          description: Błąd serwera

  /announcements:
    get:
      tags: [Messages]
      summary: Lista ogłoszeń
      responses:
        "200":
          description: OK

  /debug/messages:
    get:
      tags: [Messages]
      summary: Surowe dane z listInbox()
      parameters:
        - in: query
          name: folderId
          schema: { type: integer, default: 5 }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        "200":
          description: OK

  /homeworks:
    get:
      tags: [Homeworks]
      summary: Lista zadań domowych
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: subject
          schema: { type: string }
        - in: query
          name: includeRaw
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworksListResponse"

  /healthz:
    get:
      tags: [Misc]
      summary: Healthcheck
      responses:
        "200":
          description: OK

components:
  securitySchemes:
    XApiKey:                # klucz w nagłówku
      type: apiKey
      in: header
      name: x-api-key
    QueryApiKey:            # klucz w query
      type: apiKey
      in: query
      name: apiKey
  schemas:
    GradesResponse:
      type: object
      properties:
        ok: { type: boolean }
        count: { type: integer }
        subjects: { type: integer }
        data:
          type: array
          items:
            type: object
            properties:
              subject: { type: string }
              items:
                type: array
                items:
                  $ref: "#/components/schemas/GradeItem"
    GradeItem:
      type: object
      properties:
        subject: { type: string }
        value: { type: string, nullable: true, description: "np. +, ✓, 5" }
        area: { type: string, nullable: true }
        skill: { type: string, nullable: true }
        date: { type: string, format: date, nullable: true }
        teacher: { type: string, nullable: true }
        addedBy: { type: string, nullable: true }

    TimetableResponse:
      type: object
      properties:
        ok: { type: boolean }
        days: { type: integer }
        lessons: { type: integer }
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimetableDay"
        source: { type: string }
    TimetableDay:
      type: object
      properties:
        date: { type: string, format: date, nullable: true }
        dayName: { type: string, nullable: true, description: "np. Monday" }
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    TimetableTodayResponse:
      type: object
      properties:
        ok: { type: boolean }
        date: { type: string, format: date }
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    Lesson:
      type: object
      properties:
        start: { type: string, nullable: true, example: "08:00" }
        end: { type: string, nullable: true, example: "08:45" }
        number: { type: integer, nullable: true }
        subject: { type: string, nullable: true }
        room: { type: string, nullable: true }
        teacher: { type: string, nullable: true }
        group: { type: string, nullable: true }

    MessagesListResponse:
      type: object
      properties:
        ok: { type: boolean }
        folderId: { type: integer }
        total: { type: integer }
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageListItem"
    MessageListItem:
      type: object
      properties:
        id: { type: integer }
        folderId: { type: integer }
        subject: { type: string }
        sender: { type: string, nullable: true }
        date: { type: string, format: date-time, nullable: true }
        read: { type: boolean, nullable: true }
        hasAttachments: { type: boolean }

    MessageDetailResponse:
      type: object
      properties:
        ok: { type: boolean }
        data:
          $ref: "#/components/schemas/MessageDetail"
    MessageDetail:
      type: object
      properties:
        id: { type: integer }
        folderId: { type: integer }
        subject: { type: string }
        sender: { type: string, nullable: true }
        to:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
          nullable: true
        date: { type: string, format: date-time, nullable: true }
        body: { type: string, nullable: true }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    Attachment:
      type: object
      properties:
        name: { type: string }
        path: { type: string, nullable: true }
        size: { type: integer, nullable: true }
        mime: { type: string, nullable: true }

    HomeworksListResponse:
      type: object
      properties:
        ok: { type: boolean }
        total: { type: integer }
        data:
          type: array
          items:
            $ref: "#/components/schemas/Homework"
    Homework:
      type: object
      properties:
        id: { type: integer, nullable: true }
        subject: { type: string, nullable: true }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        assignedDate: { type: string, format: date, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        teacher: { type: string, nullable: true }